API/Alpha2/requirements.txt
Django>=3.0.7,<=3.1
djangorestframework>=3.11.0,<=3.12
psycopg2
Pillow
stripe